backend/requirements.txt
# Needed for production runtime
fastapi
gunicorn
uvicorn
SQLAlchemy
pydantic
python-jose
psycopg2-binary
passlib[bcrypt]
alembic
pydantic[email]
python-dotenv
requests
pandas
bcrypt==4.0.1